requirements.txt
aiogram==3.4.1
python-dotenv==1.0.0
aiosqlite==0.19.0
fastapi==0.110.0
uvicorn==0.29.0
jinja2==3.1.3 
python-multipart==0.0.6 
telethon==1.34.0 
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0 